fix pdb issue in konnectivity-agent helm chart
we can only set maxUnavailable or minAvailable in the spec of pdb, not
both.

In the current helm chart, we have both set always and if one of the
keys is missing in the values.yaml then the block is empty in the
rendered manifest.

This is conflicting with with the spec of pdb.
this commit refactors the helm chart so that if we have only one key
then the generated manifest then should contain only that key not the
other one.

For example: If we have `minAvailable` defined then the generated
manifest shouldn't contain `maxUnavailable`

If the user defines both key in the values.yaml then that's wrong and
I've added a note for the same that only one out of minAvailable and
maxUnavailable should be defined.

Signed-off-by: kranurag7 <anurag.kumar@syself.com>